// src/generated/models/Poce_vehiclesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Poce_vehiclespoce_status = {
  413450000: 'Available',
  413450001: 'Assigned'
} as const;
export type Poce_vehiclespoce_status = keyof typeof Poce_vehiclespoce_status;
export const Poce_vehiclesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Poce_vehiclesstatecode = keyof typeof Poce_vehiclesstatecode;
export const Poce_vehiclesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Poce_vehiclesstatuscode = keyof typeof Poce_vehiclesstatuscode;

export interface Poce_vehiclesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  poce_color?: string;
  poce_model: string;
  poce_status: Poce_vehiclespoce_status;
  poce_vehicleid: string;
  poce_vehiclename?: string;
  poce_year?: number;
  statecode: Poce_vehiclesstatecode;
  statuscode?: Poce_vehiclesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Poce_vehicles extends Poce_vehiclesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  poce_statusname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}